requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.111.0
isodate==0.6.1
